requirements-lock.txt
annotated-types==0.7.0
anyio==4.11.0
blis==1.3.0
cachetools==6.2.1
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.11
distro==1.9.0
fastapi==0.119.0
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-texttospeech==2.31.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.11.0
ko_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/ko_core_news_sm-3.8.0/ko_core_news_sm-3.8.0-py3-none-any.whl#sha256=15486b93a22ec8ee43f787390afdb4e2086851573eb12b69dd2c4f2a95b3e0a4
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.13
numpy==2.3.3
openai==2.6.0
packaging==25.0
pluggy==1.6.0
preshed==3.0.10
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
rich==14.2.0
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
smart_open==7.3.1
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.48.0
thinc==8.3.6
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.3
